fix(parser,cir): known-gaps sweep — ctor-comma declarators, VLA sizeof, const array dims
Three pre-existing gaps from claude_status.json, each fixed at the
deepest layer with a gcc/g++-verified test:

1. `Q a(1), b(2);` hung the parser: the ctor-call-syntax declaration
   path returned after `)`/`;` with no comma-continuation, leaving
   `, b(2);` orphaned for parseExprStmt to spin on. It now mirrors the
   `=`-initializer flow's convention (consume ',', re-inject a clone of
   the base-type token + storage markers). tests/testctorcomma.mad
   covers local lists, three declarators, and file-scope globals.

2. sizeof(vla_var) returned pointer size 8 — for 1D AND multidim AND
   typedef'd VLA variables (the known-gaps note was too generous; only
   sizeof(TypeName) of a runtime-sized typedef worked). Runtime dims
   are now captured into hidden __madc_vla_dim_* uint64 locals at the
   declaration (materialize_vla_dim_capture, the
   materialize_runtime_struct_size_captures pattern); the stored dim
   expressions are rewritten to read the captures, so the malloc count
   and sizeof share one evaluation (side effects once) and sizeof
   reflects the declaration-time value per C99 6.5.3.4p2 — gcc keeps
   VLA bounds in hidden temps the same way. try_parse_vla_variable_sizeof
   builds the runtime product (caps x inner dims x elem size) in the
   expression path. tests/testvlasizeof.mad (gcc-verified, including
   the n-changes-after-declaration case).

3. `const int G = 5; int g[G];` read dim 0: G's parse-time buffer is
   calloc'd and the initializer was never folded in, so
   read_constant_integer returned 0. Parse-time-known scalar-integer
   const initializers are now baked into var->data (try_eval_known_integer)
   and marked vfCONSTBAKED; the CIR read-fold admits baked consts (the
   vfCONSTDECL exclusion stays for unbaked ones), so `const int H = G+3;`
   emits a constant file-scope initializer c2mir accepts. C++ ICE
   semantics, g++-verified. Variable::set widened int->int64_t (latent
   enum-value truncation; the two (int)val casts dropped). Locals keep
   the C-canon VLA path. tests/testconstdim.mad.

Also verified: tests/testfortypedcomma.mad flake is not reproducible at
live HEAD (25 consecutive green runs) — resolved by intervening parser
work; removed from known_pre_existing_gaps.

Gates: fulltest 580/0/0/18 exit 0 both check gates GREEN (+3 new
tests); gcc.c-torture 1567 passed, failset NAME-IDENTICAL to
tmp/failset_lsq.txt (55/55; 34/21 compile/runtime split = the 87f1808
containment reclassification); SMAUG --project soak exit 124 + ready
line; zero build warnings.

+++ include/datatokens.h
@@ -97,7 +97,7 @@ public:
     inline void makeconstant() { flags |= vfCONSTANT; }
     inline bool is_global()   const { if ( (flags & vfLOCAL) && !(flags &vfSTATIC) ) return false; return true; }
     inline bool is_constant() const { if ( (flags & vfCONSTANT) ) return true; return false; }
-    bool set(int c)
+    bool set(int64_t c)
     {
 	if ( !data ) { return false; }
 	/**/ if (type == &ddCHAR)   *((char *)data) = c;
